Updated the NN_train.py such that it takes an argument that controls if the metrics are logged to wandb. Updated the data.py such that it can be ran through the terminal

+++ fagproject/src/project/NN_train.py
@@ -13,7 +13,8 @@ import os
 import importlib    
 import shutil
 
-def train_model(data_type: str, model_name: str, modul_name: str, batch_size: int, learning_rate: float, epochs: int, seed: int):
+def train_model(data_type: str, model_name: str, modul_name: str, batch_size: int, 
+                learning_rate: float, epochs: int, seed: int, log_metrics: bool):
     """
     Train our Feed forward neural networks
     """
@@ -23,33 +24,33 @@ def train_model(data_type: str, model_name: str, modul_name: str, batch_size: in
     torch.manual_seed(seed)
     torch.cuda.manual_seed(seed)
 
+    if log_metrics:
+        # Cleaning the wandb folder
+        if os.path.exists("wandb"):
+            # List all files and directories in wandb/
+            entries = os.listdir("wandb")
 
-    # Cleaning the wandb folder
-    if os.path.exists("wandb"):
-        # List all files and directories in wandb/
-        entries = os.listdir("wandb")
+        # Keep log files and 'latest-run'
+        keep_files = {"latest-run"}
+        keep_files.update([f for f in entries if f.endswith(".log")])
 
-    # Keep log files and 'latest-run'
-    keep_files = {"latest-run"}
-    keep_files.update([f for f in entries if f.endswith(".log")])
+        # Delete all other run directories
+        for entry in entries:
+            entry_path = os.path.join("wandb", entry)
+            if entry not in keep_files and entry.startswith("run-"):
+                shutil.rmtree(entry_path)
+                print(f"Deleted old run: {entry_path}")
 
-    # Delete all other run directories
-    for entry in entries:
-        entry_path = os.path.join("wandb", entry)
-        if entry not in keep_files and entry.startswith("run-"):
-            shutil.rmtree(entry_path)
-            print(f"Deleted old run: {entry_path}")
-
-    run = wandb.init(
-        entity= "lyngsberg-danmarks-tekniske-universitet-dtu",
-        project="Fagprojekt",
-        config={
-            "BATCH_SIZE": batch_size,
-            "LEARNING_RATE": learning_rate,
-            "EPOCHS": epochs,
-            "SEED": seed
-        },
-    )
+        run = wandb.init(
+            entity= "lyngsberg-danmarks-tekniske-universitet-dtu",
+            project="Fagprojekt",
+            config={
+                "BATCH_SIZE": batch_size,
+                "LEARNING_RATE": learning_rate,
+                "EPOCHS": epochs,
+                "SEED": seed
+            },
+        )
 
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
     print(f"Using device: {device}")
@@ -104,12 +105,12 @@ def train_model(data_type: str, model_name: str, modul_name: str, batch_size: in
                 outputs = model(x_batch)
                 loss = criterion(outputs, y_batch)
                 val_mse += loss.item()
-
-        wandb.log({
-            "epoch": epoch + 1,
-            "train_loss": epoch_loss,
-            "val_loss": val_mse,
-        })
+        if log_metrics:
+            wandb.log({
+                "epoch": epoch + 1,
+                "train_loss": epoch_loss,
+                "val_loss": val_mse,
+            })
 
         print(
             f"Epoch {epoch + 1}/{epochs}, "
@@ -122,29 +123,30 @@ def train_model(data_type: str, model_name: str, modul_name: str, batch_size: in
         torch.save(model.state_dict(), model_path)
         print(f"Model saved locally to {model_path}")
 
-
-        artifact = wandb.Artifact(
-            name="Neural_Network",
-            type="model",
-            description="A trained model",
-            metadata={"Val loss": val_mse,  
-                    "Date": datetime.now()},
-        )
+        if log_metrics:
+            artifact = wandb.Artifact(
+                name="Neural_Network",
+                type="model",
+                description="A trained model",
+                metadata={"Val loss": val_mse,  
+                        "Date": datetime.now()},
+            )
     else:
         model_path = os.path.join("fagproject/models/PN_models/", f"{model_name}{data_type}.pth")
         torch.save(model.state_dict(), model_path)
         print(f"Model saved locally to {model_path}")
 
-
-        artifact = wandb.Artifact(
-            name="Polynomial_Network",
-            type="model",
-            description="A trained model",
-            metadata={"Val loss": val_mse,
-                    "Date": datetime.now()},
-        )
-    artifact.add_file(model_path)
-    run.log_artifact(artifact)
+        if log_metrics:
+            artifact = wandb.Artifact(
+                name="Polynomial_Network",
+                type="model",
+                description="A trained model",
+                metadata={"Val loss": val_mse,
+                        "Date": datetime.now()},
+            )
+    if log_metrics:
+        artifact.add_file(model_path)
+        run.log_artifact(artifact)
 
 
 if __name__ == "__main__":
@@ -156,6 +158,7 @@ if __name__ == "__main__":
     parser.add_argument("--epochs", type=int, default=3)
     parser.add_argument("--seed", type=int, default=42)
     parser.add_argument("--modul_name", type=str, default="NN_models")
+    parser.add_argument("--log_metrics", type=bool, default=False)
     args = parser.parse_args()
 
-    train_model(args.data_type, args.model_name, args.modul_name, args.batch_size, args.learning_rate, args.epochs, args.seed)
+    train_model(args.data_type, args.model_name, args.modul_name, args.batch_size, args.learning_rate, args.epochs, args.seed, args.log_metrics)
